p4i integration Integrate change 67450 from 3.0 to main. Description for change 67450:
: Change 67450 by porten@porten-komet on 2002/07/04 15:24:13
: 
: 	honour Motif-style Help menu position
: 	silenced some warnings
: 
: Affected files ...
: 
: ... //depot/qt/3.0/examples/chart/chartform.cpp#12 edit
: ... //depot/qt/3.0/examples/chart/chartform_canvas.cpp#4 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 67451]

+++ examples/chart/chartform_canvas.cpp
@@ -61,7 +61,7 @@ void ChartForm::drawPieChart( const double scales[], double total, int )
 {
     double width = m_canvas->width();
     double height = m_canvas->height();
-    double size = width > height ? height : width;
+    int size = int(width > height ? height : width);
     int x = int(width / 2);
     int y = int(height / 2);
     int angle = 0;
